ODX-module-restructured: Create a parent module for the child modules
