--- modules/custom/odp_core/config/install/system.action.odp_vbo_unpublished_action.yml
@@ -1,7 +1,0 @@
-langcode: en
-status: true
-id: odp_vbo_unpublished_action
-label: 'Unpublish latest node'
-type: node
-plugin: "odp_vbo_unpublished_action"
-configuration: {  }
