--- modules/custom/odp_core/odp_core.info.yml
@@ -1,9 +1,0 @@
-name: Opendevx Core
-description: Customizing Opendevx core
-type: module
-core: 8.x
-core_version_requirement: ^8 || ^9
-
-dependencies:
-  - domain:domain
-  - views_bulk_operations:views_bulk_operations
