--- modules/custom/odp_core/src/Plugin/views/filter/ProgramIdFilterPlugin.php
@@ -1,54 +1,0 @@
-<?php
-
-namespace Drupal\odp_core\Plugin\views\filter;
-
-use Drupal\views\Plugin\views\filter\FilterPluginBase;
-use Symfony\Component\DependencyInjection\ContainerInterface;
-use Drupal\odp_core\Program\ProgramDomainInterface;
-
-/**
- * Filters program on current domain.
- *
- * @ingroup views_filter_handlers
- *
- * @ViewsFilter("program_id_filter")
- */
-class ProgramIdFilterPlugin extends FilterPluginBase {
-
-  /**
-   * Object ProgramDomain.
-   *
-   * @var Drupal\odp_core\Program\ProgramDomainInterface
-   */
-  protected $programDomain;
-
-  /**
-   * Pass the dependency to the object constructor.
-   */
-  public function __construct(array $configuration, $plugin_id, $plugin_definition, ProgramDomainInterface $program_domain) {
-    parent::__construct($configuration, $plugin_id, $plugin_definition);
-    $this->programDomain = $program_domain;
-  }
-
-  /**
-   * {@inheritdoc}
-   */
-  public static function create(ContainerInterface $container, array $configuration, $plugin_id, $plugin_definition) {
-    return new static(
-      $configuration,
-      $plugin_id,
-      $plugin_definition,
-      $container->get('odp_core.program_domain'),
-    );
-  }
-
-  /**
-   * {@inheritdoc}
-   */
-  public function query() {
-    if ($program_id = $this->programDomain->getProgramDomainId()) {
-      $this->query->addWhere('and', 'groups_field_data.id', $program_id);
-    }
-  }
-
-}
