+++ modules/custom/odp_program/modules/odp_domain/odp_domain.info.yml
@@ -1,0 +1,10 @@
+name: OpenDevPortal Domain
+description: Provides the Domain Customisations.
+type: module
+package: ODP
+core: 8.x
+core_version_requirement: ^8 || ^9
+
+dependencies:
+  - domain:domain
+  - views_bulk_operations:views_bulk_operations
