+++ modules/custom/odp_program/modules/odp_domain/odp_domain.routing.yml
@@ -1,0 +1,7 @@
+odp_domain.add_program_page:
+  path: '/program/{pid}/content/create'
+  defaults:
+    _controller: '\Drupal\odp_domain\Controller\ProgramsContentController::addPage'
+    _title: 'Add Content'
+  requirements:
+    _role: admin
