+++ modules/custom/odp_program/modules/odp_domain/src/Plugin/Action/OpenDevPortalVboPublished.php
@@ -1,0 +1,39 @@
+<?php
+
+namespace Drupal\odp_domain\Plugin\Action;
+
+use Drupal\views_bulk_operations\Action\ViewsBulkOperationsActionBase;
+use Drupal\Core\Session\AccountInterface;
+
+/**
+ * Publishes an entity.
+ *
+ * @Action(
+ *   id = "odp_vbo_published_action",
+ *   label = @Translation("Publish latest node"),
+ *   type = "node",
+ *   confirm = TRUE,
+ * )
+ */
+class OpenDevPortalVboPublished extends ViewsBulkOperationsActionBase {
+
+  /**
+   * {@inheritdoc}
+   */
+  public function execute($entity = NULL) {
+    $entity->setPublished()->save();
+  }
+
+  /**
+   * {@inheritdoc}
+   */
+  public function access($object, AccountInterface $account = NULL, $return_as_object = FALSE) {
+    if ($object->getEntityType() === 'node') {
+      $access = $object->access('update', $account, TRUE)
+        ->andIf($object->status->access('edit', $account, TRUE));
+      return $return_as_object ? $access : $access->isAllowed();
+    }
+    return TRUE;
+  }
+
+}
