+++ modules/custom/odp_taxonomy/modules/odp_organisation/odp_organisation.info.yml
@@ -1,4 +1,4 @@
-name: Developer Portal Organisation
+name: OpenDevPortal Organisation
 description : 'Extend the organisation taxonomy term for developer portal.'
 type: module
 package: ODP
