+++ modules/custom/odp_user/odp_user.info.yml
@@ -5,4 +5,4 @@ package: ODP
 core: 8.x
 core_version_requirement: ^8 || ^9
 dependencies:
-  - odp_core:odp_core
+  - odp_domain:odp_domain
