+++ modules/custom/odp_user/odp_user.services.yml
@@ -1,7 +1,7 @@
 services:
   odp_user.organisation:
    class: Drupal\odp_user\Organisation
-   arguments: ['@tempstore.private', '@database', '@current_user', '@entity_type.manager', '@path.alias_manager', '@odp_core.program_domain']
+   arguments: ['@tempstore.private', '@database', '@current_user', '@entity_type.manager', '@path.alias_manager', '@odp_domain.program_domain']
 
   odp_user.logger:
    class: Drupal\odp_user\Logger\Logger
