+++ opendevportal.info.yml
@@ -166,7 +166,7 @@ install:
  - odp_breadcrumb
  - odp_content
  - odp_content_reference
- - odp_core
+ - odp_domain
  - odp_fence_tab
  - odp_menu
  - odp_navigation_block
@@ -174,6 +174,7 @@ install:
  - odp_notification
  - odp_organisation
  - odp_paragraph
+ - odp_program
  - odp_rapidoc_formatter
  - odp_sdk
  - odp_subscription
